requirements.txt
Flask==3.0.3
python-docx==1.1.2
pypdf==4.2.0
reportlab==4.2.2
pytest==8.3.2
